requirements.txt
numpy==1.22.4
pandas==1.4.2
pickleshare==0.7.5
Pillow==9.1.1
plotly==5.8.0
scikit-learn==1.1.1
scipy==1.9.0
sklearn==0.0
streamlit==1.10.0
